@@ -1,9 +1,9 @@
-import { makeStyles } from '@material-ui/core'
+import { makeStyles } from '@mui/material'
 
 export const useStyles = makeStyles((theme) => ({
   deleteButton: {
-    color: theme.status.blood,
-  },
+    color: theme.status.blood
+  }
 }))
 
 export default useStyles
