@@ -1,13 +1,13 @@
-import { makeStyles } from '@material-ui/core'
+import { makeStyles } from '@mui/material'
 
 export const useStyles = makeStyles((theme) => ({
   deleteButton: {
-    color: theme.status.blood,
+    color: theme.status.blood
   },
   bottomButtons: {
-    display: "flex",
-    justifyContent: "space-between",
-  },
+    display: 'flex',
+    justifyContent: 'space-between'
+  }
 }))
 
 export default useStyles
