@@ -1,9 +1,9 @@
-import { makeStyles } from '@material-ui/core'
+import { makeStyles } from '@mui/material'
 
 export const useStyles = makeStyles((theme) => ({
   minusButton: {
     color: theme.palette.secondary.main
-  },
+  }
 }))
 
 export default useStyles
