@@ -1,7 +1,8 @@
-import { Box } from '@material-ui/core'
+import { Box } from '@mui/material'
 import React from 'react'
 
 interface TabPanelProps {
+  children?: React.ReactNode
   value: number
   index: number
 }
