@@ -1,15 +1,15 @@
-import { makeStyles } from '@material-ui/core'
+import { makeStyles } from '@mui/material'
 
 export const useStyles = makeStyles((theme) => ({
   taperedRule: {
-    display: "block",
-    width: "100%",
-    height: "7px",
-    border: "none",
+    display: 'block',
+    width: '100%',
+    height: '7px',
+    border: 'none',
     color: theme.status.blood,
     fill: theme.status.blood,
-    margin: "0.5em 0",
-  },
+    margin: '0.5em 0'
+  }
 }))
 
 export default useStyles
