@@ -1,10 +1,10 @@
-import { createMuiTheme } from '@material-ui/core'
+import { createMuiTheme } from '@mui/material/styles'
 
-declare module "@material-ui/core/styles/createMuiTheme" {
+declare module '@mui/material/styles/createMuiTheme' {
   interface Theme {
     status: {
-      blood: React.CSSProperties["color"]
-      light: React.CSSProperties["color"]
+      blood: React.CSSProperties['color']
+      light: React.CSSProperties['color']
     }
     custom: {
       boxShadow: string
@@ -12,8 +12,8 @@ declare module "@material-ui/core/styles/createMuiTheme" {
   }
   interface ThemeOptions {
     status: {
-      blood: React.CSSProperties["color"]
-      light: React.CSSProperties["color"]
+      blood: React.CSSProperties['color']
+      light: React.CSSProperties['color']
     }
     custom: {
       boxShadow: string
@@ -23,65 +23,65 @@ declare module "@material-ui/core/styles/createMuiTheme" {
 
 const theme = createMuiTheme({
   custom: {
-    boxShadow: "0.1rem 0 0.5rem #afaba5, -0.1rem 0 0.5rem #afaba5",
+    boxShadow: '0.1rem 0 0.5rem #afaba5, -0.1rem 0 0.5rem #afaba5'
   },
   typography: {
-    fontSize: 16,
+    fontSize: 16
   },
   status: {
-    blood: "#922610",
-    light: "#f6efdc",
+    blood: '#922610',
+    light: '#f6efdc'
   },
   palette: {
     primary: {
-      main: "#eee3c5",
-      dark: "#d3c7a6",
+      main: '#eee3c5',
+      dark: '#d3c7a6'
     },
     secondary: {
-      main: "#c77b09",
-    },
+      main: '#c77b09'
+    }
   },
   overrides: {
     MuiFormControl: {
       root: {
-        width: "100%",
-      },
+        width: '100%'
+      }
     },
     MuiTextField: {
       root: {
-        "& > label": {
-          fontSize: "1.5rem",
-          fontWeight: "bold",
-          color: "#1b1b1b",
-          whiteSpace: "nowrap",
+        '& > label': {
+          fontSize: '1.5rem',
+          fontWeight: 'bold',
+          color: '#1b1b1b',
+          whiteSpace: 'nowrap'
         },
-        "& > label.Mui-focused.Mui-focused": {
-          color: "#c77b09",
+        '& > label.Mui-focused.Mui-focused': {
+          color: '#c77b09'
         },
-        "& > .MuiInput-underline:after": {
-          borderBottom: `2px solid #c77b09`,
+        '& > .MuiInput-underline:after': {
+          borderBottom: `2px solid #c77b09`
         },
-        "& > div": {
-          marginTop: "24px",
-        },
-      },
+        '& > div': {
+          marginTop: '24px'
+        }
+      }
     },
     MuiTabs: {
       indicator: {
-        height: "6px",
-      },
+        height: '6px'
+      }
     },
     MuiTab: {
       wrapper: {
-        fontWeight: "bold",
-      },
+        fontWeight: 'bold'
+      }
     },
     MuiButton: {
       label: {
-        fontWeight: "bold",
-      },
-    },
-  },
+        fontWeight: 'bold'
+      }
+    }
+  }
 })
 
 export default theme
