@@ -5,13 +5,13 @@ import { itemState } from 'recoil/atoms'
 
 import useStyles from './ItemStats.styles'
 
-const DescriptionBlock: React.FC = (props) => {
+const DescriptionBlock: React.FC<{ children?: React.ReactNode }> = (props) => {
   const { children } = props
   const classes = useStyles()
   return <p className={classes.blockDescription}>{children}</p>
 }
 
-const MainDescription: React.FC = (props) => {
+const MainDescription: React.FC<{ children?: React.ReactNode }> = (props) => {
   const { children } = props
   const classes = useStyles()
   return <div className={classes.mainDescription}>{children}</div>
@@ -33,19 +33,19 @@ export const ItemStats: React.FC = () => {
             return (
               <div className={classes.featureContainer} key={key}>
                 <h3 className={classes.featureName}>{feature.featureName}</h3>
-                {feature.featureDescription.split("\n").map((value, key) => {
-                return (
-                  <DescriptionBlock key={`description-${key}`}>
-                    {value}
-                  </DescriptionBlock>
-                )
-              })}
+                {feature.featureDescription.split('\n').map((value, key) => {
+                  return (
+                    <DescriptionBlock key={`description-${key}`}>
+                      {value}
+                    </DescriptionBlock>
+                  )
+                })}
               </div>
             )
           })}
           {currentItem.mainDescription && (
             <MainDescription>
-              {currentItem.mainDescription.split("\n").map((value, key) => {
+              {currentItem.mainDescription.split('\n').map((value, key) => {
                 return (
                   <DescriptionBlock key={`description-${key}`}>
                     {value}
