@@ -5,11 +5,12 @@ import React from 'react'
 import useStyles from './StatsContainer.styles'
 
 interface StatsContainerProps {
+  children?: React.ReactNode
   className?: string
 }
 
 export const StatsContainer: React.FC<StatsContainerProps> = (props) => {
-  const { children, className = "" } = props
+  const { children, className = '' } = props
   const classes = useStyles()
   const divStyle = {
     backgroundImage: 'url(' + statblockparch + ')'
@@ -19,9 +20,7 @@ export const StatsContainer: React.FC<StatsContainerProps> = (props) => {
     <div className={`${classes.root} ${className} stats-container`}>
       <OrangeBorder />
       <div className={`${classes.content} stats-background`} style={divStyle}>
-        <div className={classes.margins}>
-        {children}
-        </div>
+        <div className={classes.margins}>{children}</div>
       </div>
       <OrangeBorder />
     </div>
